+++ packages/runed/src/lib/utilities/MediaQuery/index.ts
@@ -1,0 +1,1 @@
+export { MediaQuery } from "./MediaQuery.svelte.js";