+++ sites/docs/src/lib/components/demos/debounced.svelte
@@ -1,21 +1,19 @@
 <script lang="ts">
 	import { Debounced } from "runed";
+	import DemoContainer from "$lib/components/demo-container.svelte";
+	import Input from "$lib/components/ui/input/input.svelte";
 
 	let search = $state("");
 	const debounced = new Debounced(() => search, 500);
 </script>
 
-<div class="rounded-md bg-card p-8">
-	<input
-		class="w-full rounded-md border border-input bg-background px-3 py-2 text-sm ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none disabled:opacity-50"
-		bind:value={search}
-		placeholder="Search the best utilities for Svelte 5"
-	/>
+<DemoContainer class="flex flex-col gap-4">
+	<Input bind:value={search} placeholder="Search the best utilities for Svelte 5" />
 	<p>
 		{#if debounced.current}
-			You searched for: <b>{debounced.current}</b>
+			You searched for: <b class="text-brand">{debounced.current}</b>
 		{:else}
 			Search for something above!
 		{/if}
 	</p>
-</div>
+</DemoContainer>